feat(modularization): S-13a -- extract overview + activity to web/modules/overview.js
Moves the Overview page (loadOverview, formatRelative, fmtTokensShort,
ovActivityIcon, _ovRenderActivityFeed) and the Activity live-poll
(startActivityPoll, stopActivityPoll, loadActivity, renderActivity,
initActivity) out of app.js into web/modules/overview.js.

The module version adds a _ovLoadGen generation guard against concurrent
loadOverview() races, which was absent in the inline app.js version.

app.js imports the five public symbols; the inline Activity IIFE click
handler is replaced by initActivity({ openTerminalModal }).

Smoke: 0 pageerror, overview renders (7/7 agents, kpiTasks visible).
vitest: 2873 tests green.
